'Any' object to properly encode types better
Any object can hold a reference/pointer to any known type or pass
small-ish types by value.  It can be made to do type conversions to base
types and remembers const-ness of references.  This allows better
behavior when setting member variable values or passing parameters to
member functions, as well as return values from getters or member
functions.  It could be simplified and cleaned up a bit, as well as
tested a bit more thoroughly.